PageFeedService, more test and some refactoring
