+++ src/test/java/org/ominidi/facebook/service/PageFeedServiceTest.java
@@ -1,5 +1,6 @@
 package org.ominidi.facebook.service;
 
+import com.restfb.types.Post;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
@@ -7,6 +8,7 @@ import org.mockito.runners.MockitoJUnitRunner;
 import org.ominidi.facebook.repository.Feed;
 
 import static org.mockito.Mockito.*;
+import static org.junit.Assert.*;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PageFeedServiceTest {
@@ -20,5 +22,19 @@ public class PageFeedServiceTest {
         service.getFeedPosts();
 
         verify(feed).getConnection();
+    }
+
+    @Test
+    public void shouldReturnASinglePost() {
+        Long id = 221946658231380L;
+        Post post = new Post();
+        post.setId(id.toString());
+        when(feed.getObject(id)).thenReturn(post);
+
+        FeedAware service = new PageFeedService(feed);
+        Post result = service.getPostById(id);
+
+        verify(feed).getObject(id);
+        assertEquals(id.toString(), result.getId());
     }
 }
